nextjs extractions
