+++ examples/counter/web-nextjs/src/app/page.tsx
@@ -3,78 +3,18 @@
 import type { NextPage } from "next";
 import { useEffect, useState } from "react";
 
-import init_core, { process_event, handle_response, view } from "shared/shared";
-import type {
-  Event,
-  HttpResponse,
-  SseResponseVariantChunk,
-  SseResponseVariantDone,
-} from "shared_types/types/shared_types";
+import init_core from "shared/shared";
 import {
-  EffectVariantRender,
   ViewModel,
-  EffectVariantHttp,
-  EffectVariantServerSentEvents,
   EventVariantStartWatch,
   EventVariantDecrement,
   EventVariantIncrement,
-  Request,
 } from "shared_types/types/shared_types";
-import {
-  BincodeSerializer,
-  BincodeDeserializer,
-} from "shared_types/bincode/mod";
 
-import { request as http } from "./http";
-import { request as sse } from "./sse";
-
-type Response = HttpResponse | SseResponseVariantChunk | SseResponseVariantDone;
+import { update } from "./core";
 
 const Home: NextPage = () => {
   const [state, setState] = useState(new ViewModel("", false));
-
-  function dispatch(event: Event) {
-    const serializer = new BincodeSerializer();
-    event.serialize(serializer);
-    const effects = process_event(serializer.getBytes());
-    processEffects(effects);
-  }
-
-  function respond(uuid: number[], response: Response) {
-    const serializer = new BincodeSerializer();
-    response.serialize(serializer);
-    const effects = handle_response(
-      new Uint8Array(uuid),
-      serializer.getBytes()
-    );
-    processEffects(effects);
-  }
-
-  async function processEffects(effects: Uint8Array) {
-    const requests = deserializeRequests(effects);
-
-    for (const { uuid, effect } of requests) {
-      switch (effect.constructor) {
-        case EffectVariantRender: {
-          setState(deserializeView(view()));
-          break;
-        }
-        case EffectVariantHttp: {
-          const request = (effect as EffectVariantHttp).value;
-          const response = await http(request);
-          respond(uuid, response);
-          break;
-        }
-        case EffectVariantServerSentEvents: {
-          const request = (effect as EffectVariantServerSentEvents).value;
-          for await (const response of sse(request)) {
-            respond(uuid, response);
-          }
-          break;
-        }
-      }
-    }
-  }
 
   useEffect(
     () => {
@@ -82,7 +22,7 @@ const Home: NextPage = () => {
         await init_core();
 
         // Initial event
-        dispatch(new EventVariantStartWatch());
+        update(new EventVariantStartWatch(), setState);
       }
 
       loadCore();
@@ -102,13 +42,13 @@ const Home: NextPage = () => {
         <div className="buttons section is-centered">
           <button
             className="button is-primary is-warning"
-            onClick={() => dispatch(new EventVariantDecrement())}
+            onClick={() => update(new EventVariantDecrement(), setState)}
           >
             {"Decrement"}
           </button>
           <button
             className="button is-primary is-danger"
-            onClick={() => dispatch(new EventVariantIncrement())}
+            onClick={() => update(new EventVariantIncrement(), setState)}
           >
             {"Increment"}
           </button>
@@ -117,20 +57,5 @@ const Home: NextPage = () => {
     </main>
   );
 };
-
-function deserializeRequests(bytes: Uint8Array) {
-  const deserializer = new BincodeDeserializer(bytes);
-  const len = deserializer.deserializeLen();
-  const requests: Request[] = [];
-  for (let i = 0; i < len; i++) {
-    const request = Request.deserialize(deserializer);
-    requests.push(request);
-  }
-  return requests;
-}
-
-function deserializeView(bytes: Uint8Array) {
-  return ViewModel.deserialize(new BincodeDeserializer(bytes));
-}
 
 export default Home;
