+++ examples/simple_counter/web-nextjs/src/app/page.tsx
@@ -4,43 +4,18 @@ import type { NextPage } from "next";
 import Head from "next/head";
 import { useEffect, useState } from "react";
 
-import init_core, { process_event, view } from "shared/shared";
-import type { Event } from "shared_types/types/shared_types";
+import init_core from "shared/shared";
 import {
-  Request,
   ViewModel,
-  EffectVariantRender,
   EventVariantReset,
   EventVariantIncrement,
   EventVariantDecrement,
 } from "shared_types/types/shared_types";
-import {
-  BincodeDeserializer,
-  BincodeSerializer,
-} from "shared_types/bincode/mod";
+
+import { update } from "./core";
 
 const Home: NextPage = () => {
   const [state, setState] = useState(new ViewModel("0"));
-
-  function dispatch(event: Event) {
-    const serializer = new BincodeSerializer();
-    event.serialize(serializer);
-    const effects = process_event(serializer.getBytes());
-    processEffects(effects);
-  }
-
-  async function processEffects(bytes: Uint8Array) {
-    const requests = deserializeRequests(bytes);
-
-    for (const { uuid: _, effect } of requests) {
-      switch (effect.constructor) {
-        case EffectVariantRender: {
-          setState(deserializeView(view()));
-          break;
-        }
-      }
-    }
-  }
 
   useEffect(
     () => {
@@ -48,7 +23,7 @@ const Home: NextPage = () => {
         await init_core();
 
         // Initial event
-        dispatch(new EventVariantReset());
+        update(new EventVariantReset(), setState);
       }
 
       loadCore();
@@ -69,19 +44,19 @@ const Home: NextPage = () => {
           <div className="buttons section is-centered">
             <button
               className="button is-primary is-danger"
-              onClick={() => dispatch(new EventVariantReset())}
+              onClick={() => update(new EventVariantReset(), setState)}
             >
               {"Reset"}
             </button>
             <button
               className="button is-primary is-success"
-              onClick={() => dispatch(new EventVariantIncrement())}
+              onClick={() => update(new EventVariantIncrement(), setState)}
             >
               {"Increment"}
             </button>
             <button
               className="button is-primary is-warning"
-              onClick={() => dispatch(new EventVariantDecrement())}
+              onClick={() => update(new EventVariantDecrement(), setState)}
             >
               {"Decrement"}
             </button>
@@ -91,20 +66,5 @@ const Home: NextPage = () => {
     </>
   );
 };
-
-function deserializeRequests(bytes: Uint8Array) {
-  const deserializer = new BincodeDeserializer(bytes);
-  const len = deserializer.deserializeLen();
-  const requests: Request[] = [];
-  for (let i = 0; i < len; i++) {
-    const request = Request.deserialize(deserializer);
-    requests.push(request);
-  }
-  return requests;
-}
-
-function deserializeView(bytes: Uint8Array) {
-  return ViewModel.deserialize(new BincodeDeserializer(bytes));
-}
 
 export default Home;
